pyproject.toml
[tool.poetry]
name = "orders-management-system"
version = "0.1.0"
description = ""
authors = ['Aleksey']
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2024.12.1"
crispy-bootstrap5 = "^2024.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
mypy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
ignore_missing_imports = true